+++ src/email_builder/mod.rs
@@ -7,7 +7,8 @@ mod sender;
 
 pub use bulk_email::BulkEmail;
 pub use email::Email;
-pub use message::{Message, MessageTemplate, Reader};
+pub use message::Message;
+pub use message::MessageTemplate;
 pub use mime::MimeFormat;
 pub use receiver::Receiver;
 pub use sender::Sender;
