+++ src/email_builder/sender.rs
@@ -6,7 +6,7 @@ use crate::arg;
 pub struct Sender;
 
 impl Sender {
-    pub fn new<'a>(matches: &'a ArgMatches) -> Result<&'a str, anyhow::Error> {
+    pub fn init<'a>(matches: &'a ArgMatches) -> Result<&'a str, anyhow::Error> {
         if matches.is_present(arg::SENDER) {
             match matches.value_of(arg::SENDER) {
                 Some(sender) => Ok(sender),
